42 mini talk

+++ libft/ft_atoi.c
@@ -6,7 +6,7 @@
 /*   By: adardour <adardour@student.1337.fr>        +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/10/08 21:47:05 by adardour          #+#    #+#             */
-/*   Updated: 2022/10/20 13:35:23 by adardour         ###   ########.fr       */
+/*   Updated: 2022/10/29 21:05:02 by adardour         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -21,9 +21,9 @@ static int	check(char c)
 
 int	ft_atoi(const char *str)
 {
-	int			i;
-	int			sign;
-	long int	result;
+	int	i;
+	int	sign;
+	int	result;
 
 	i = 0;
 	sign = 1;
